refactor(agent): group device components
