+++ src/containers/Devices/components/DevicesEditItemList.js
