+++ src/containers/Devices/components/Map.js
@@ -5,9 +5,9 @@ import L from 'leaflet'
 delete L.Icon.Default.prototype._getIconUrl
 
 L.Icon.Default.mergeOptions({
-  iconRetinaUrl: require('../../node_modules/leaflet/dist/images/marker-icon-2x.png'),
-  iconUrl: require('../../node_modules/leaflet/dist/images/marker-icon.png'),
-  shadowUrl: require('../../node_modules/leaflet/dist/images/marker-shadow.png')
+  iconRetinaUrl: require('../../../../node_modules/leaflet/dist/images/marker-icon-2x.png'),
+  iconUrl: require('../../../../node_modules/leaflet/dist/images/marker-icon.png'),
+  shadowUrl: require('../../../../node_modules/leaflet/dist/images/marker-shadow.png')
 })
 
 class Map extends Component {
