+++ src/containers/Devices/components/Sections/Applications.js
@@ -2,9 +2,9 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import ReactWinJS from 'react-winjs'
 import WinJS from 'winjs'
-import EmptyMessage from '../../../components/EmptyMessage'
-import ContentPane from '../../../components/ContentPane'
-import Loader from '../../../components/Loader'
+import EmptyMessage from '../../../../components/EmptyMessage'
+import ContentPane from '../../../../components/ContentPane'
+import Loader from '../../../../components/Loader'
 
 class Applications extends Component {
