+++ src/containers/Devices/components/Sections/GeolocationList.js
