+++ src/containers/Devices/components/Sections/index.js
