+++ src/containers/Invitations/InvitationsPage.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import InvitationsPendingPage from './InvitationsPendingPage'
-import Enroll from '../Devices/Enroll'
+import Enroll from '../Devices/components/Enroll'
 import EmptyMessage from '../../components/EmptyMessage'
 
 export default class InvitationsPage extends Component {
